Extract editor playback source helpers
